ml-engine/requirements.txt
# Deep Learning
torch>=2.1.0
torch-geometric>=2.4.0
torchvision>=0.16.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Graph Neural Networks
networkx>=3.2.1
dgl>=1.1.0

# NLP (for regulatory document processing)
transformers>=4.35.0
huggingface-hub>=0.19.0
sentencepiece>=0.1.99

# Blockchain Integration
web3>=6.11.0
eth-account>=0.10.0
eth-utils>=2.3.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# API and Database
fastapi>=0.104.0
uvicorn>=0.24.0
pymongo>=4.6.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Logging and Monitoring
loguru>=0.7.2
prometheus-client>=0.19.0

# IPFS Integration
ipfshttpclient>=0.8.0a2
